int as payload size

+++ protocol/include/fractal.h
@@ -742,7 +742,7 @@ struct RTPPacket {
 	int id;
 	short index;
 	short num_indices;
-	short payload_size;
+	int payload_size;
 	bool is_a_nack;
 	// data at the end of the struct, in the case of a truncated packet
 	uint8_t data[MAX_PAYLOAD_SIZE];
